fix message type

+++ include/encrypt.h
@@ -19,8 +19,8 @@ void FinalRound(unsigned char * state, unsigned char * key);
 void AESEncrypt(uint32* message, unsigned char * expandedKey, uint8* encryptedMessage);
 void KeyExpansionCore(unsigned char * in, unsigned char i);
 void KeyExpansion(unsigned char inputKey[16], unsigned char expandedKeys[176]);
-void addPadding(const uint32* message , int messageLen , uint32* paddedMessage);
-void startEncryption(const uint32* message , uint32 messageLen, uint8* macPtr, uint32* macLengthPtr);
+void addPadding(const uint8* message , int messageLen , uint32* paddedMessage);
+void startEncryption(const uint8* message , uint32 messageLen, uint8* macPtr, uint32* macLengthPtr);
 
 // #endif
 #endif  // ENCRYPTION_ENCRYPT_H_
