+++ source/main.c
@@ -6,7 +6,7 @@
 int main(void)
 {
     SecOC_MainFunctionTx();
-    uint32 data[16] = {'h','o','s','a' , 'm'};
+    uint8 data[16] = {'h','o','s','a' , 'm'};
     uint8 mac[16];
     uint32 macLen = 8;
     startEncryption(data , 8 , mac ,&macLen);
